+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 
 ## Beta 3
-?
+**Breaking Changes**
+* modified Services.Geoprocessing constructor to account for changes in esri leaflet core at [Release Candidate 5](https://github.com/Esri/esri-leaflet/blob/master/CHANGELOG.md#release-candidate-5)
+* added a generic setParam() method to replace previous setters
+
+**Changes**
+* in addition to L.GeoJSON geometries, L.LatLng, L.LatLngBounds, and L.Marker are now considered valid GP inputs
 
 ## Beta 2